[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.03 ms, Memory: 78.1 MB -BaekjoonHub

+++ 프로그래머스/lv1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -1,12 +1,13 @@
 class Solution {
     public long[] solution(long x, int n) {
-        int s;
+            int s;
             try{
-                 s = (int) (((x * n) / x) < 0 ? ((x * n) / x) * -1 : ((x * n) / x));
+                s = (int) (((x * n) / x) < 0 ? ((x * n) / x) * -1 : ((x * n) / x));
             }catch (ArithmeticException e){
                 s = n;
             }
-            long[] answer = new long[(int) s];
+
+            long[] answer = new long[s];
             for (int i = 0; i < answer.length; i++) {
                 answer[i] = (long) (i + 1) * x;
             }
